Introduce FaStackDirective and refactor FaStackComponent to use host directives
